[ENH] Remove `CenteredGrid` serialization to simplify `Grid` data model
- Eliminated `CenteredGrid` serialization and deserialization logic, including related binary metadata and byte handling.
- Refactored `name_id_map` assignment for improved element data formatting.
- Standardized float representations in test cases for consistent output.

+++ test/test_modules/test_geophysics/test_gravity.py
@@ -67,7 +67,7 @@ def test_gravity():
     gp.set_centered_grid(
         grid=geo_model.grid,
         centers=np.array([[6, 0, 4]], dtype="float"),
-        resolution=np.array([10, 10, 100], dtype="float"),
+        resolution=np.array([10, 10, 100]),
         radius=np.array([16000, 16000, 16000], dtype="float")  # ? This radius makes 0 sense but it is the original one in gempy v2
     )
